Move wave_propagator to Wave to allow its use by other propagators
