+++ spyro/solvers/time_integration_central_difference.py
@@ -5,7 +5,7 @@ from . import helpers
 from .. import utils
 
 
-def central_difference(wave, source_id=0):
+def central_difference(wave):
     """
     Perform central difference time integration for wave propagation.
 
@@ -13,8 +13,6 @@ def central_difference(wave, source_id=0):
     -----------
     wave: Spyro object
         The Wave object containing the necessary data and parameters.
-    source_id: int (optional)
-        The ID of the source being propagated. Defaults to 0.
 
     Returns:
     --------
@@ -22,8 +20,10 @@ def central_difference(wave, source_id=0):
             A tuple containing the forward solution and the receiver output.
     """
     if wave.sources is not None:
-        wave.sources.current_source = source_id
+        source_id = wave.sources.current_source
         rhs_forcing = fire.Function(wave.function_space)
+    else:
+        source_id = 0
 
     filename, file_extension = wave.forward_output_file.split(".")
     output_filename = filename + "sn" + str(source_id) + "." + file_extension
